projects/flaskView/requirements.txt
alembic==1.7.7
anyio==3.5.0
appdirs==1.4.4
async-generator==1.10
bcrypt==3.1.7
cryptography==36.0.2
cssselect2==0.4.1
distlib==0.3.4
dnspython==2.2.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
fonttools==4.27.1
html5lib==1.1
httpcore==0.13.6
httpx==0.18.2
idna==2.9
immutables==0.16
importlib-metadata==3.4.0
importlib-resources==5.4.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
jsonpickle==2.0.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
monotonic==1.5
mysql-connector-python==8.0.25
netifaces==0.10.6
openpyxl==3.0.4
pycparser==2.21
pydyf==0.1.2
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyphen==0.11.0
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2022.1
requests==2.27.1
SQLAlchemy==1.3.16
Werkzeug==2.0.3
WTForms==2.3.1
zipp==3.6.0
zopfli==0.1.9
